Copy data out of those vorbis/theora internal buffers
This is way safer when we're doing drawing and such on a separate
thread from the decoder, since it can overwrite those buffers when
working on the next frame!

Unfortunately if the drawing can't keep pace with the decoding this
will queue frames until we run out of memory. Need a further fix to
discard late frames.