Support local filesystem playback
File URLs are now recognized and sent to a separate implementation of
the base OGVInputStream (now an abstract class) for local data loading.
This should be suitable for files that have been locally downloaded
manually, or are included in the application bundle.

i/o may happen on the decode thread, but should stay out of the main
thread. Future improvements may better separate the i/o.

Other URL types are sent to OGVHTTPInputStream, which is pretty much
as the older OGVInputStream version and supports HTTP(S) only.

Example is changed to bundle the local copy of the VP8 switching test
file and a short low-res version of the 'Big Buck Bunny' intro (ogg)
to confirm that both WebM and Ogg files play back this way.

Fixes https://github.com/brion/OGVKit/issues/88
Fixes https://github.com/brion/OGVKit/issues/155