AVFoundation-backed playback for MP4
Will allow pulling in MP3 and MP4 files for transcoding
once encoding is added.

Only works with URL inputs for now.

Changed OGVAudioFeeder to use a circular buffer internally,
and to stop when starved instead of trying to pad with silence.
That was causing confusion on the audio clock with the MP4
sample.